MSVC warning fix

+++ tests/json_test/json_test.cpp
@@ -6019,12 +6019,12 @@ suite complex_test = [] {
     4
   ]
 ])";
-      std::vector<std::complex<int16_t>> cx{};
+      std::vector<std::complex<float>> cx{};
       expect(!glz::read_json(cx, s));
-      expect(cx[0].real() == 1);
-      expect(cx[0].imag() == 2);
-      expect(cx[1].real() == 3);
-      expect(cx[1].imag() == 4);
+      expect(cx[0].real() == 1.f);
+      expect(cx[0].imag() == 2.f);
+      expect(cx[1].real() == 3.f);
+      expect(cx[1].imag() == 4.f);
    };
 };
